Downloads types of authors for bills

Downloads types of authors for bills

cham_typeauthors_bills(ascii = T)

Arguments

ascii
logical. If TRUE, certain strings are converted to ascii format.

Value

A tibble, of classes tbl_df, tbl and data.frame.

Note

Requesting data from a long period of time with details = TRUE will return a large object in terms of memory. It will also be rather unwieldy, with many columns.

Examples

cham_typeauthors_bills()

#> # A tibble: 43 x 2
#>    type_id                           type_desc
#>      <chr>                               <chr>
#>  1       1                   Comissao Diretora
#>  2      10                   Grupo de Trabalho
#>  3    1000 Importacao das proposicoes inativas
#>  4   10000                          Deputado  
#>  5      11                            Conselho
#>  6     110    Comissao da Camara dos Deputados
#>  7      12            Procuradoria Parlamentar
#>  8   12000       Orgao da Camara dos Deputados
#>  9      13                        Corregedoria
#> 10   13000           Comissao Temporaria da CD
#> # ... with 33 more rows