Downloads votes of a specific bill by providing type, number and year

Downloads votes of a specific bill by providing type, number and year. A bill can have more than one roll call, and the API does not provide an id to identify them So we provide one (rollcall_id).

cham_votes(type, number, year, ascii = TRUE)

Arguments

type
character. The type of the bill. For example, "PL" for law proposal ("projeto de lei"), "PEC" for constitutional amendments ("projeto de emenda constitucional"), "PDC" for legislative decree ("decreto legislativo"), and "PLP" for supplementary laws ("projeto de lei complementar).

number
integer. The number of the bill.

year
integer. The year of the bill.

ascii
logical. If TRUE, certain strings are converted to ascii format.

Value

A tibble, of classes tbl_df, tbl and data.frame.

Note

Requesting data from a long period of time with details = TRUE will return a large object in terms of memory. It will also be rather unwieldy, with many columns.

Examples

cham_votes(type = "PL", number = "1992", year = "2007")

#> # A tibble: 2,052 x 31
#>    rollcall_id decision_summary decision_date decision_time rollcall_subject
#>    <chr>       <chr>            <chr>         <chr>         <chr>           
#>  1 PL-1992-20… Aprovada a Sube… 28/2/2012     20:26         SUBEMENDA SUBST…
#>  2 PL-1992-20… Aprovada a Sube… 28/2/2012     20:26         SUBEMENDA SUBST…
#>  3 PL-1992-20… Aprovada a Sube… 28/2/2012     20:26         SUBEMENDA SUBST…
#>  4 PL-1992-20… Aprovada a Sube… 28/2/2012     20:26         SUBEMENDA SUBST…
#>  5 PL-1992-20… Aprovada a Sube… 28/2/2012     20:26         SUBEMENDA SUBST…
#>  6 PL-1992-20… Aprovada a Sube… 28/2/2012     20:26         SUBEMENDA SUBST…
#>  7 PL-1992-20… Aprovada a Sube… 28/2/2012     20:26         SUBEMENDA SUBST…
#>  8 PL-1992-20… Aprovada a Sube… 28/2/2012     20:26         SUBEMENDA SUBST…
#>  9 PL-1992-20… Aprovada a Sube… 28/2/2012     20:26         SUBEMENDA SUBST…
#> 10 PL-1992-20… Aprovada a Sube… 28/2/2012     20:26         SUBEMENDA SUBST…
#> # … with 2,042 more rows, and 26 more variables: session_id <chr>,
#> #   DEM_orientation <chr>, GOV_orientation <chr>, Minoria_orientation <chr>,
#> #   PDT_orientation <chr>, PMDB_orientation <chr>, PMN_orientation <chr>,
#> #   PP_orientation <chr>, PRB_orientation <chr>,
#> #   PrPtdobPrpPhsPtcPslPrtb_orientation <chr>, PsbPtbPcdob_orientation <chr>,
#> #   PSC_orientation <chr>, PSD_orientation <chr>, PSDB_orientation <chr>,
#> #   PSOL_orientation <chr>, PT_orientation <chr>, PvPps_orientation <chr>,
#> #   legislator_id <chr>, legislator_name <chr>, legislator_state <chr>,
#> #   legislator_party <chr>, legislator_vote <chr>, ReprPMN_orientation <chr>,
#> #   type_bill <chr>, number_bill <chr>, year_bill <chr>