Downloads and tidies data on the agenda in the Federal Senate

Downloads and tidies data on the agenda in the Federal Senate.

sen_agenda(initial_date = NULL, end_date = NULL, house = NULL,
  supervisory = NULL, legislator = NULL, details = FALSE, ascii = TRUE)

Arguments

initial_date
(character) start date of the period requested. This parameter must be in the format YYYYMMDD (Year-Month-Day). A value for this parameter is necessary, all others are optional.

end_date
(character) final date for period requested. Format YYYYMMDD.

house
(character). The acronym for the legislative house for which results are requested. Options are SF (Senado Federal, Federal Senate), CN (Congresso Nacional, National Congress - joint meeting of the Senate and Chamber), and CA Camara dos Deputados, Chamber of Deputies.

supervisory
character. Name of the commission or supervisory body. A data frame of these can be seen with data("commissions").

legislator
integer. The numeric code given to each senator. A dataframe with these values is returned from the sen_senator_list() function.

details
(logical). If details is equal to TRUE, the data returned is an expanded dataset with additional details. This is not recommended unless necessary.

ascii
logical. If TRUE, certain strings are converted to ascii format.

Value

A tibble, of classes tbl_df, tbl and data.frame.

Note

Requesting data from a long period of time with details = TRUE will return a large object in terms of memory. It will also be rather unwieldy, with many columns.

Examples

# NOT RUN {
sen_agenda(initial_date = "20161105", end_date = "20161125")
sen_agenda(initial_date = "20161105", end_date = "20161125",
legislator = 4988)
# }