Downloads and tidies information on the possible locations a piece of legislation can currently be passing through the Federal Senate

Downloads and tidies information on the possible locations a piece of legislation can currently be passing through the Federal Senate.

sen_bills_locations(active = NULL, ascii = TRUE)

Arguments

active
character. Options are "Yes", "No" or NULL, the default. "Yes" returns active subtypes, "No" returns inactive subtypes, while the default returns both.

ascii
logical. If TRUE, strips Latin characters from strings.

Value

A tibble, of classes tbl_df, tbl and data.frame.

Examples

sen_bills_locations()

#> # A tibble: 100 x 7
#>    loc_id loc_abbr loc_name loc_type loc_type_descr loc_house
#>    <chr>  <chr>    <chr>    <chr>    <chr>          <chr>    
#>  1 3      ADVOSF   Advocac… A        Unidade Admin… Senado F…
#>  2 434    ATRSGM   Assesso… A        Unidade Admin… Senado F…
#>  3 38     CAE      Comissa… C        Colegiado      Senado F…
#>  4 40     CAS      Comissa… C        Colegiado      Senado F…
#>  5 449    CCAI     Comissa… C        Colegiado      Congress…
#>  6 2147   CCC      Comissa… C        Colegiado      Senado F…
#>  7 34     CCJ      Comissa… C        Colegiado      Senado F…
#>  8 767    CCS      Conselh… C        Colegiado      Congress…
#>  9 1363   CCT      Comissa… C        Colegiado      Senado F…
#> 10 336    CD       Camara … A        Unidade Admin… Camara d…
#> # … with 90 more rows, and 1 more variable: loc_date_created <dttm>