Downloads and tidies information on bills that have been recently updated in the Federal Senate

Downloads and tidies information on bills that have been recently updated in the Federal Senate.

sen_bills_updates(update = NULL, year = NULL, number = NULL,
  type = NULL, days = 5, ascii = TRUE)

Arguments

update
character. This is the type of update that can be applied to a bill. For a dataframe of these, use the sen_bills_update_types() function, and the variable update_name that is returned.

year
character. Year of the bill, if a specific bill is requested. Format YYYY.

number
bill number.

type
type of legislation.

days
integer. The number of days to consider when requesting information on recent updates. The maximum is 30 and the default is 5.

ascii
logical. If TRUE, strips Latin characters from strings.

Value

A tibble, of classes tbl_df, tbl and data.frame.

Examples

# Bills from 2014 that have had a "despacho" update in the last 15 days, if
# they exist:

# NOT RUN {
desp_2014 <- sen_bills_updates(update = "Despacho", year = 2014, days = 15)

# PLS bills that have been updated in the last 10 days, if they exist:
pls <- sen_bills_updates(type = "PLS", days = 10)
# }