Returns results from the plenary in the Federal Senate for a specified date

Returns results from the plenary in the Federal Senate for a specified date.

sen_plenary_result(date = NULL, ascii = TRUE)

Arguments

date
character. Format YYYYMMDD.

ascii
logical. If TRUE, the default, strips Latin characters from the results.

Value

A tibble, of classes tbl_df, tbl and data.frame.

Examples

# get info for the 3rd of March 2016:
x <- sen_plenary_result(date = "20160303")
# Earlier periods may not have information:

sen_plenary_result(date = "20110405")

#> # A tibble: 16 x 17
#>    bill_id bill  bill_type bill_number bill_year session_id session_number
#>    <chr>   <chr> <chr>     <chr>       <chr>     <chr>      <chr>         
#>  1 98237   "PRO… MPV       00510       2010      2070       42            
#>  2 87281   "PRO… PLC       00141       2008      2070       42            
#>  3 85711   "PRO… PDS       00116       2008      2070       42            
#>  4 99140   "PRO… PDS       00079       2011      2070       42            
#>  5 99483   "REQ… REQ       00246       2011      2070       42            
#>  6 99514   "REQ… RQS       00248       2011      2070       42            
#>  7 99504   "REQ… REQ       00249       2011      2070       42            
#>  8 99506   "REQ… REQ       00250       2011      2070       42            
#>  9 99507   "REQ… REQ       00251       2011      2070       42            
#> 10 99505   "REQ… REQ       00252       2011      2070       42            
#> 11 99503   "REQ… REQ       00253       2011      2070       42            
#> 12 99532   "REQ… REQ       00255       2011      2070       42            
#> 13 99533   "REQ… REQ       00256       2011      2070       42            
#> 14 99573   "REQ… REQ       00270       2011      2070       42            
#> 15 99574   "REQ… REQ       00271       2011      2070       42            
#> 16 99641   "REQ… REQ       00305       2011      2070       42            
#> # ... with 10 more variables: session_date <dttm>, session_time <chr>,
#> #   session_type <chr>, session_type_abbr <chr>, session_house <chr>,
#> #   bill_house <chr>, bill_report <chr>, bill_details <chr>, bill_result <chr>,
#> #   bill_sponsor <chr>