Downloads and tidies information on the commissions on which senators have served or are serving in the Federal Senate

Downloads and tidies information on the commissions on which senators have served or are serving in the Federal Senate.

sen_senator_commissions(id = 0, ascii = TRUE)

Arguments

id
integer. This number represents the id of the senator you wish to get information on. These ids can be extracted from the API using the sen_senator_list() function, where they will appear as the first column in the data frame returned, under the name 'id'.

ascii
logical. If TRUE, certain strings are converted to ascii format.

Value

A tibble, of classes tbl_df, tbl and data.frame.

See also

sen_senator_list()

Examples

Armando <- sen_senator_commissions(id = 715)

#> Error in sen_senator_commissions(id = 715): No data match your request. Have you checked the 'id' argument?