Nominal votes in the Brazilian Federal Senate

This is a dataset of the nominal votes in the Brazilian Federal Senate, from all those available on the API from 1991 onwards.

senate_nominal_votes

Format

A data frame with 60691 rows and 8 variables

Value

vote_date: POSIXct, date the vote took place.

bill_id: id of the bill in the Senate API database.

bill: bill type, year and number.

legislature: legislature number.

senator_id: unique id of the senator.

senator_name: the senator's name.

senator_vote: vote cast. 1 = "yes"; 0 = "no", NA = other.

senator_party: political party the senator was in when the vote took place.

senator_state: state the senator represented when the vote took place.

Note

These data can easily be grouped by legislature if so desired, using the legislature variable.