Downloads and tidies data for lists of bills in Brazilian Chamber of Deputies

Downloads and tidies data for lists of bills in Brazilian Chamber of Deputies

cham_bills(type = "", number = "", year = "", initial_date = "",
  end_date = "", part_name_author = "", id_type_author = "",
  abbreviation_party_author = "", abbreviation_st_author = "",
  gender_author = "", cod_state = "", cod_branch_state = "",
  still = "", ascii = T)

Arguments

type
character. The type of the bill. For example, "PL".

number
integer. The number of the bill

year
integer. The year of the bill.

initial_date
(character) start date of the period requested. This parameter must be in the format YYYY-MM-DD (Year-Month-Day).

end_date
(character) final date for period requested. Format YYYYMMDD.

part_name_author
(character). Filter bills by a part of the author's name

id_type_author
(character). Filter bills by type of a branch_author for the bill.

abbreviation_party_author
(character). Filter bills by the party of the author

abbreviation_st_author
(character). Filter bills by the state of the author

gender_author
(character). If "M", bills proposed by male authors are filtered. If "F", female authors. If empty, every bill is included.

cod_state
(integer). Filter bills by a part of the author's name

cod_branch_state
(integer). Filter by the id of the branch regarding the bill.

still
(integer). If 1, only bills that are still being discussed are included. If 2, only ended bills. If empty, every bill is included.

ascii
logical. If TRUE, certain strings are converted to ascii format.

Value

A tibble, of classes tbl_df, tbl and data.frame.

Examples

cham_bills(type = "PL", year = 2011, number = 2718, initial_date="2011-16-11")

#> # A tibble: 1 x 31
#>   bill_id bill_name type_id type_abr type_name branch_id branch_abr branch_name
#>   <chr>   <chr>     <chr>   <chr>    <chr>     <chr>     <chr>      <chr>      
#> 1 527478  PL 2718/… 139     PL       Projeto … 180       "PLEN    … PLENARIO   
#> # … with 23 more variables: presentation_date <chr>, ementa_txt <chr>,
#> #   ementa_explanation <chr>, procedure_cod <chr>, procedure_txt <chr>,
#> #   consideration_id <chr>, consideration_txt <chr>, author1_name <chr>,
#> #   author1_id <chr>, author1_party_cod <chr>, author1_party_abr <chr>,
#> #   author1_state <chr>, number_authors <chr>, last_action <chr>,
#> #   last_action_txt <chr>, situation_id <chr>, situation_desc <chr>,
#> #   situation_branch_cod <chr>, situation_branch_abr <chr>,
#> #   main_bill_cod <chr>, main_bill <chr>, gender_id <chr>,
#> #   num_branches_state <chr>