Downloads and tidies data on the agenda in the Federal Senate

Downloads and tidies data on the agenda in the Federal Senate.

sen_agenda(initial_date = NULL, end_date = NULL, house = NULL,
  supervisory = NULL, legislator = NULL, details = FALSE,
  ascii = TRUE)

Arguments

initial_date
(character) start date of the period requested. This parameter must be in the format YYYYMMDD (Year-Month-Day). A value for this parameter is necessary, all others are optional.

end_date
(character) final date for period requested. Format YYYYMMDD.

house
(character). The acronym for the legislative house for which results are requested. Options are SF (Senado Federal, Federal Senate), CN (Congresso Nacional, National Congress - joint meeting of the Senate and Chamber), and CA Camara dos Deputados, Chamber of Deputies.

supervisory
character. Name of the commission or supervisory body. A data frame of these can be seen with data("commissions").

legislator
integer. The numeric code given to each senator. A dataframe with these values is returned from the sen_senator_list() function.

details
(logical). If details is equal to TRUE, the data returned is an expanded dataset with additional details. This is not recommended unless necessary.

ascii
logical. If TRUE, certain strings are converted to ascii format.

Value

A tibble, of classes tbl_df, tbl and data.frame.

Note

Requesting data from a long period of time with details = TRUE will return a large object in terms of memory. It will also be rather unwieldy, with many columns.

Examples

sen_agenda(initial_date = "20161105", end_date = "20161125")

#> # A tibble: 87 x 12
#>    agenda_id agenda_title agenda_name agenda_type agenda_date        
#>    <chr>     <chr>        <chr>       <chr>       <dttm>             
#>  1 5551      CCS, as 09h… Reuniao de… Reuniao     2016-11-07 00:00:00
#>  2 5537      CDH, as 10h… Audiencia … Extraordin… 2016-11-07 00:00:00
#>  3 5553      CCS, as 14h… Reuniao de… Reuniao     2016-11-07 00:00:00
#>  4 5576      CJD, as 14h… Audiencia … Reuniao     2016-11-07 00:00:00
#>  5 5591      CDH, as 15h… Audiencia … Extraordin… 2016-11-07 00:00:00
#>  6 5563      CCT, as 08h… Audiencia … Extraordin… 2016-11-08 00:00:00
#>  7 5535      CDH, as 09h… Audiencia … Extraordin… 2016-11-08 00:00:00
#>  8 5578      CMA, as 09h… Deliberati… Extraordin… 2016-11-08 00:00:00
#>  9 5598      CCJ,CAE, as… Audiencia … Conjunta    2016-11-08 00:00:00
#> 10 5581      CTG, as 11h… NA          Extraordin… 2016-11-08 00:00:00
#> # ... with 77 more rows, and 7 more variables: agenda_time <chr>,
#> #   agenda_status <chr>, agenda_place <chr>, agenda_commission_house <chr>,
#> #   agenda_commission_abbr <chr>, agenda_commission_id <chr>,
#> #   agenda_commission_meeting_number <chr>

sen_agenda(initial_date = "20161105", end_date = "20161125",
legislator = 4988)

#> # A tibble: 87 x 15
#>    agenda_id agenda_title agenda_name agenda_type agenda_date        
#>    <chr>     <chr>        <chr>       <chr>       <dttm>             
#>  1 5551      CCS, as 09h… Reuniao de… Reuniao     2016-11-07 00:00:00
#>  2 5537      CDH, as 10h… Audiencia … Extraordin… 2016-11-07 00:00:00
#>  3 5553      CCS, as 14h… Reuniao de… Reuniao     2016-11-07 00:00:00
#>  4 5576      CJD, as 14h… Audiencia … Reuniao     2016-11-07 00:00:00
#>  5 5591      CDH, as 15h… Audiencia … Extraordin… 2016-11-07 00:00:00
#>  6 5563      CCT, as 08h… Audiencia … Extraordin… 2016-11-08 00:00:00
#>  7 5535      CDH, as 09h… Audiencia … Extraordin… 2016-11-08 00:00:00
#>  8 5578      CMA, as 09h… Deliberati… Extraordin… 2016-11-08 00:00:00
#>  9 5598      CCJ,CAE, as… Audiencia … Conjunta    2016-11-08 00:00:00
#> 10 5581      CTG, as 11h… NA          Extraordin… 2016-11-08 00:00:00
#> # ... with 77 more rows, and 10 more variables: agenda_time <chr>,
#> #   agenda_status <chr>, agenda_place <chr>, agenda_commission_house <chr>,
#> #   agenda_commission_abbr <chr>, agenda_commission_id <chr>,
#> #   agenda_commission_meeting_number <chr>, legislator_name <chr>,
#> #   agenda_commission_member <chr>, agenda_commission_relator <chr>