Search for data on legislation in the Brazilian Federal Senate

Search for data on legislation in the Brazilian Federal Senate.

sen_bill_search(year = "", year_law = "", topic_id = "",
  situation_id = "", date_presented_init = "", date_presented_end = "",
  date_situation_init = "", date_situation_end = "", complementary = "",
  present = "", rapporteur = "", author = "", number = "", type = "",
  law_number = "", party_abbr_author = "", author_state = "",
  in_passage = "", ascii = TRUE)

Arguments

year
integer. Four-digit year, such as 2013.

year_law
integer. Year of introduction of the law, such as 2013.

topic_id
character. For a data frame of topic ids and their meanings, use sen_bills_topics().

situation_id
character.

date_presented_init
character. Date when the bill that you're searching for was first presented. In the format YYYYMMDD.

date_presented_end
character. See above.

date_situation_init
character. See above.

date_situation_end
character. See above.

complementary
character. Either blank, "Yes", or "No".

present
character. Either blank, "Yes", or "No"; an indicator for whether the bill is current or not.

rapporteur
character. Name of the rapporteur of the bill, if known.

author
character. Author/sponsor of the bill. For a list, use sen_bill_sponsors().

number
character. Bill number, if known.

type
character. Bill type. For a data frame of possible bill types, run sen_bills_types().

law_number
character. Number of the law resulting from the bill, if known.

party_abbr_author
character. The short text code for the party of the author of the bill. For a list of the parties, use sen_parties().

author_state
character. The state of the senator. For a full list, see UF().

in_passage
character. Either blank, "Yes", or "No".

ascii
logical. If TRUE, certain strings are converted to ascii format.

Value

A tibble, of classes tbl_df, tbl and data.frame.

Examples

# search for legislation from 2014:

# NOT RUN {
two14 <- sen_bill_search(year = 2014)
# }